SOCINT-259 Splitting respository tests into class specific ones
