--- src/main/java/uk/gov/ons/ctp/integration/rhsvc/repository/impl/DatastoreStartupCheckData.java
@@ -1,13 +1,0 @@
-package uk.gov.ons.ctp.integration.rhsvc.repository.impl;
-
-import lombok.AllArgsConstructor;
-import lombok.Data;
-import lombok.NoArgsConstructor;
-
-@Data
-@NoArgsConstructor
-@AllArgsConstructor
-class DatastoreStartupCheckData {
-  private String timestamp;
-  private String hostname;
-}
