+++ src/test/java/uk/gov/ons/ctp/integration/rhsvc/event/impl/SurveyUpdateEventReceiverImplUnit_test.java
@@ -13,7 +13,7 @@ import org.mockito.junit.jupiter.MockitoExtension;
 import uk.gov.ons.ctp.common.FixtureHelper;
 import uk.gov.ons.ctp.common.error.CTPException;
 import uk.gov.ons.ctp.common.event.model.SurveyUpdateEvent;
-import uk.gov.ons.ctp.integration.rhsvc.repository.impl.RespondentSurveyRepository;
+import uk.gov.ons.ctp.integration.rhsvc.repository.RespondentSurveyRepository;
 
 @ExtendWith(MockitoExtension.class)
 public class SurveyUpdateEventReceiverImplUnit_test {
