+++ src/test/resources/uk/gov/ons/ctp/integration/rhsvc/repository/PackageFixture.CollectionExercise.json
